refactor: package uuid and logging utils
